0.19.4 new package for Hair + refactoring
