+++ wfrp/src/main/java/com/warhammer/wfrp/hair/models/HighElfHairColor.java
@@ -1,4 +1,4 @@
-package com.warhammer.wfrp.model.hair;
+package com.warhammer.wfrp.hair.models;
 
 import jakarta.persistence.*;
 import lombok.Getter;
