+++ wfrp/src/main/java/com/warhammer/wfrp/hair/repositories/DwarfHairColorRepository.java
@@ -1,6 +1,6 @@
-package com.warhammer.wfrp.repository.hair;
+package com.warhammer.wfrp.hair.repositories;
 
-import com.warhammer.wfrp.model.hair.DwarfHairColor;
+import com.warhammer.wfrp.hair.models.DwarfHairColor;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 public interface DwarfHairColorRepository extends JpaRepository<DwarfHairColor, Integer> {
