+++ wfrp/src/main/java/com/warhammer/wfrp/hair/repositories/HalflingHairColorRepository.java
@@ -1,6 +1,6 @@
-package com.warhammer.wfrp.repository.hair;
+package com.warhammer.wfrp.hair.repositories;
 
-import com.warhammer.wfrp.model.hair.HalflingHairColor;
+import com.warhammer.wfrp.hair.models.HalflingHairColor;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 public interface HalflingHairColorRepository extends JpaRepository<HalflingHairColor, Integer> {
