+++ wfrp/src/main/java/com/warhammer/wfrp/hair/repositories/HumanHairColorRepository.java
@@ -1,6 +1,6 @@
-package com.warhammer.wfrp.repository.hair;
+package com.warhammer.wfrp.hair.repositories;
 
-import com.warhammer.wfrp.model.hair.HumanHairColor;
+import com.warhammer.wfrp.hair.models.HumanHairColor;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 public interface HumanHairColorRepository extends JpaRepository<HumanHairColor, Integer> {
